+++ src/main/java/com/umc/product/global/exception/constant/Domain.java
@@ -28,5 +28,6 @@ public enum Domain {
     FIGMA,
     LLM,
     ANALYTICS,
-    MAINTENANCE
+    MAINTENANCE,
+    FEEDBACK
 }
